Revert "deposit: Remove 'parent' deposit"
This reverts commit f6905cdf547c8a1325fe8b3a0cbf6cb2373031a6.

That commit was a first step toward making loaders write releases
instead of revisions.

Unfortunately, we will still write revisions for a non-negligeable time,
so I prefer to defer the removal of parent deposit revisions to the
moment we actually make that switch, so we don't end up with inconsistent
revisions.