debian: Fix confusion between the two versions
'version' was documented as the intrinsic version (eg. '0.7.2-3') and
'full_version' as the one containing the suite name (eg. 'stretch/contrib/0.7.2-3').

In practice, it was the opposite, except in a few incorrect test.

This commit fixes said tests, and renamed 'full_version' to
'intrinsic_version'.

This is only a refactoring, the behavior is unchanged for now;
but a future commit will remove the 'version' (which is extrinsic) from
the release name (which should contain only data intrinsic to the DSC).