loader/opam/tests: Do not run actual opam init command call
Instead, this fixes the pseudo fake opam root present in tests datadir folder to the
right structure and then installs it properly for the test to use.

Another benefit from this is now inhibiting any actual network happening during the opam
init cli call.

This also reverts the previous commit which added a --yes in the opam init cli.
This was not working and not the right solution in the end.

Related to T3973