loader: Ensure success is False when entering exception handler
The post_load method of a loader can raise an exception so we must
ensure to turn back the success variable to False in that case.

For instance, the subversion loader post_load checks that latest
exported revision is consistent with what the official subversion
client produces. If it is not an exception will be raised to set
the visit status to partial.