package/archive: Handle tarball artifact with null time
An artifact without time info can be provided in the artifacts list
parameter of the loader (for instance last modification date
is not available for tarballs coming from github releases).

That case was not handled by the archive loader wich was resulting
in loading error so add fix for it.