{Content|Directory}Loader: Adapt support for checksums
This adapts the content/directory loader implementations to use directly a checksums
dict which is now sent by the listers.

This improves the loader to check those checksums when retrieving the artifact (content
or tarball). Thanks to a bump in the swh.model version, this is now able to deal with
sha512 checksums checks as well.

This also aligns with the current package loaders which now are also checking the
integrity of the tarballs they ingest.

Related to T3781