rpm: Fix package versions sorting
Release 22.0 of packaging module can no longer parse invalid Python version
number, an exception is now raised.

RPM loader used the keys of the packages dict as version numbers to sort,
which are in the form "<distribution>/<edition>/<package_version_number>",
but those cannot be parsed anymore.

So use intrinsic version numbers of packages instead to sort the list of
versions.