Fix presentation of the loader specification table
After switching to the PyData Sphinx theme, the very large table
with the loader specification became too wide to be readable.
These changes make the table scrollable again, remove the right sidebar,
highlight the name column and add stripe to rows.

They are stopgap measure as such a table might not the best way to
present this information.