debian: Handle invalid version string when parsing changelog
Some packages can have an invalid version string in their changelog
which raises a ValueError when attempting to parse it in the
get_intrinsic_package_metadata method of the loader.

As a consequence each package release with such bogus entry in the
changelog was discarded from the snapshot created by the loader.

So prefer to get the raw string version instead of parsing it to
workaround that issue.

Fix #1493.