Allow centralized configuration for loaders (with optional override)
That should allow centralized configuration for loaders without convolutions in the
sysadm dimension. That is generically adding options that may not be understood by all
loaders without making them fail. The ones supporting those options will use it. The
others not supporting those will just not use it.

This is an alternative implementation of loosening the base loader constructor to accept
kwargs which could pose side-effecty problems.

Refs. swh/infra/ci-cd/swh-charts!62