Revert to ArchiveLoader and allow it to pass through task checks
As it's already running in production, the 'tar' visit type is now immutable. So we
cannot change anything related to it.

So, instead rename it as before and adapt the check utility functions to allow those
checks to be bypassed for some specific cases (like the ArchiveLoader). New loaders
should fix their assertion failure immediately though.

Refs. swh/infra/sysadm-environment#4906