opam: Use proper way to list all versions of a package
opam 2.1 stores packages metadata in a tarball instead of having its
content unpacked as with previous opam versions, so the workaround of
walking the inner directories of the opam root to get all versions of
a package is no longer working with opam 2.1.

So use proper way to get all versions of a package by calling the
following command:

$ opam show --root <opam_root> <package> --field all-versions

The previous issue reported in TODO comment was due to the fact that
new opam repositories were not added in the default opam switch, this
is now fixed and packages metadata from all repositories can be displayed
using the opam show command.