rpm: Ensure correct package versions ordering
Previously package versions were sorted according to packages dict
keys but this is not reliable as older versions can be sorted after
newer ones.

Prefer to sort package versions according to their build time then
as it produces a correct ordering and ensure the HEAD branch alias
will target the most recent version of a package.