nar: Exclude only a single vcs type when computing hash
When computing the recursive nar hash of a directory fetched from
a VCS like git or svn, only special directories related to the used
VCS (.git or .svn) should be excluded. For instance when a directory
was fetched from git, only .git folders should be excluded.

Related to swh/devel/swh-loader-git#4751.