package/nixguix: Adapt tests data to ensure all tarballs are different
This is required after the changes introduced in 54d8c7d and it is also
more consistent with real data provided by nix or guix.

Moreover, this loader should be removed soon as a better lister have been
implemented for nixguix that no longer create loading tasks with nixguix
visit type.