package/utils: Set accept-encoding to identity for download request
requests package set the value of the Accept-Encoding HTTP header to
"gzip, deflate" by default and some servers (https://download.ocamlcore.org/
for instance) will then send a compressed version of the artifact to download
with response header content-encoding usually set to gzip.

Nevertheless, this conflicts with the code checking if the response bytes
should be uncompressed as it should not when Content-Encoding header is equal
to gzip and Content-Type is equal to application/*gzip.

As artifacts to download are usually tarballs already compressed with gzip,
set the Accept-Encoding request header to identity in order to force the
server to send raw artifact bytes without the Content-Encoding header set.