package: Do not report NotFound to sentry when package info is missing
It exist cases when a package version is referenced in package versions
list but package info for that version is missing.

This type of error is not fixable from the SWH side so there is no need
to report these errors to sentry as it generates a lot of noise in the
sentry dashboards.

Fixes #4751.