loader: Add support to load a content or a directory from a NAR archive
Source of NixOS packages produced by the nixguix lister targets NAR archives
downloadable from the NixOS HTTP cache. Those archives can contain files and
directories but also a single file (typically a tarball or a patch file).

As checksums to verify must be computed on the content of such archive and
not on the NAR file, an extra step is required to unpack its content.