loader: Optimize NodeLoader for already archived content or directory
When the checksums parameter is provided to NodeLoader, check if
associated extids are available in the archive and avoid to download
and process the content or directory to archive again as it was already
done in a previous load. Simply fetch the content or directory from the
archive instead.

Fixes #4756.